Initial design renderings
Input from product team: that's a logo. The eyes at the at the bottom are the eyes of the cat. I put them there because the eyes are hard to see. (I call them cute eyes!). She has decided on "Dorlu". And she's provided some artistic direction on the logo. It is a cupcake on the left.  And a butterfly on the right, cat at the bottom.  All in a heart. Dev notes: The reference image at the very bottom is a close up of the cats eyes which are rendered small to scale, but have detail and depth. Particular emphasis on cute eye motif.